smishguard-backend/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
Django==4.1.4
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
filelock==3.8.2
fonttools==4.38.0
future==0.18.2
httplib2==0.21.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.2
nltk==3.7
numpy==1.23.5
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
pip==22.2.2
platformdirs==2.5.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-whois==0.8.0
pytz==2022.6
regex==2022.10.31
requests==2.28.1
scikit-learn==1.0.2
scipy==1.9.3
setuptools==65.5.0
six==1.16.0
sklearn==0.0.post1
soupsieve==2.3.2.post1
sqlparse==0.4.3
threadpoolctl==3.1.0
tqdm==4.64.1
uritools==4.0.0
urlextract==1.7.1
urllib3==1.26.13
wheel==0.37.1
xgboost==0.90
